% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_to_code.R
\name{doc_to_code}
\alias{doc_to_code}
\title{Write documentation for code}
\usage{
doc_to_code(file, level = "light")
}
\arguments{
\item{file}{Path to the code file}

\item{level}{Documentation level of detail. Defaults to light. Light: Provides a description of functions and pipes. Heavy: All lines are documented (not very useful).}
}
\value{
A new R file is created adding a documentation to the code
}
\description{
Write documentation for code
}
\examples{
code_to_doc("my_code.R", "light")
code_to_doc("my_code.R", "heavy")
}
