% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droxygen.R
\name{droxygen}
\alias{droxygen}
\title{Create roxygen doc from function}
\usage{
droxygen(file)
}
\arguments{
\item{file}{Path to the input code file with the function}
}
\value{
A new file is created as a roxygen doc.  The file ends with "_droxygend" and is saved in the same directory as the input file.
}
\description{
Create roxygen doc from function
}
