% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droxygen.R
\name{droxygen}
\alias{droxygen}
\title{Write documentation for code}
\usage{
droxygen(file)
}
\arguments{
\item{file}{Path to the code file with the function}
}
\value{
A new R file is created adding a 10 word documentation comment to the code
}
\description{
Write documentation for code
}
\examples{
droxygen("my_new_package_function.R")
}
