% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droxygen.R
\name{droxygen}
\alias{droxygen}
\title{Create roxygen doc from function}
\usage{
droxygen(file)
}
\arguments{
\item{file}{Path to the code file with the function}
}
\value{
A new file is created as a roxygen doc
}
\description{
Create roxygen doc from function
}
\examples{
droxygen("my_new_package_function.R")
}
